LayoutTests/fast/mediastream/MediaStream-add-remove-tracks-expected.txt

Test adding and removing tracks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS stream1.getAudioTracks().length is 1
PASS stream1.getVideoTracks().length is 1
PASS stream2.getAudioTracks().length is 1
PASS stream2.getVideoTracks().length is 1
PASS stream1.getAudioTracks().length is 0
PASS stream1.getVideoTracks().length is 0
PASS stream1.getTrackById(audioTrack.id) is null
PASS stream1.getTrackById(videoTrack.id) is null
PASS stream1.getAudioTracks().length is 1
PASS stream1.getVideoTracks().length is 1
PASS stream1.getAudioTracks()[0].kind is audioTrack.kind
PASS stream1.getAudioTracks()[0].label is audioTrack.label
PASS stream1.getVideoTracks()[0].kind is videoTrack.kind
PASS stream1.getVideoTracks()[0].label is videoTrack.label
PASS stream1.getAudioTracks().length is 1
PASS stream1.getVideoTracks().length is 1
PASS stream1.getAudioTracks().length is 2
PASS stream1.getVideoTracks().length is 2
PASS stream1.addTrack(audioTrack) threw exception Error: InvalidStateError: DOM Exception 11.
PASS stream1.removeTrack(audioTrack) threw exception Error: InvalidStateError: DOM Exception 11.
PASS successfullyParsed is true

TEST COMPLETE
